Move test to v3 API
